move chains info
